+++ apps/backend/src/routers/public-metamcp.ts
@@ -2,7 +2,7 @@ import cors from "cors";
 import express from "express";
 
 import { endpointsRepository } from "../db/repositories/endpoints.repo";
-import openApiRouter from "./public-metamcp/openapi";
+import { openApiRouter } from "./public-metamcp/openapi";
 import sseRouter from "./public-metamcp/sse";
 import streamableHttpRouter from "./public-metamcp/streamable-http";
