+++ apps/backend/src/routers/public-metamcp/openapi/types.ts
@@ -1,0 +1,18 @@
+import { ApiKeyAuthenticatedRequest } from "../../../middleware/api-key-auth.middleware";
+
+export interface ToolExecutionRequest extends ApiKeyAuthenticatedRequest {
+  params: { tool_name: string };
+}
+
+export interface OpenApiSchema {
+  openapi: string;
+  info: {
+    title: string;
+    description: string;
+    version: string;
+  };
+  paths: Record<string, unknown>;
+  components: {
+    schemas: Record<string, unknown>;
+  };
+}
